Added league drop-down
New script `005` gets the current leagues and saves them to a txt file. GUI now reads that txt and incorporates it into a drop-down menu allowing a user to select a league.  I have NOT integrated this with the other scripts yet, so this drop-down is not completely functional yet.